Refactor checko code to cache runs on git-hash and input, and reuse SSH connections when performing git actions
This drastically speeds up the overall execution time of evaluating projects, as well as simplifies job management.

A highlight feature is that jobs are no longer stored in the DB, and then repopulated in the hub, but rather runs are cached. This means that when we load checko, the hub remains empty, and then we queue all runs again. These runs then checks the DB before spawning any jobs, and uses these results if applicable. Thus, we get to remove a lot of the code that managed storing, retriving, and repopulating runs as jobs, for a simpler model that also does more finegrained caching.

+++ driver/src/lib.rs
@@ -85,7 +85,7 @@ impl<M: Debug + Send + Sync + 'static> Driver<M> {
     where
         M: Clone,
     {
-        self.config.compile.as_ref().map(|compile| {
+        if let Some(compile) = &self.config.compile {
             if let Some(job) = self.current_compilation.write().unwrap().take() {
                 job.kill();
             }
@@ -98,20 +98,25 @@ impl<M: Debug + Send + Sync + 'static> Driver<M> {
                 .write()
                 .unwrap()
                 .replace(job.clone());
-            tokio::spawn({
-                let driver = self.clone();
-                let job = job.clone();
-                async move {
-                    tracing::debug!("waiting for it to compile...");
-                    let state = job.wait().await;
-                    tracing::debug!(?state, "finished!");
-                    if let JobState::Succeeded = state {
-                        *driver.latest_successfull_compile.write().unwrap() = Some(job)
+            tokio::spawn(
+                {
+                    let driver = self.clone();
+                    let job = job.clone();
+                    async move {
+                        tracing::debug!("waiting for it to compile...");
+                        let state = job.wait().await;
+                        tracing::debug!(?state, "finished!");
+                        if let JobState::Succeeded = state {
+                            *driver.latest_successfull_compile.write().unwrap() = Some(job)
+                        }
                     }
                 }
-            });
-            job
-        })
+                .in_current_span(),
+            );
+            Some(job)
+        } else {
+            None
+        }
     }
     pub fn config(&self) -> &RunOption {
         &self.config
